Import of Templates and folder structure
Alright, this is the base from which we will work from.

Reset.css is all the reset css code (em duh...)
style.css will be all the desktop and general styling for all of the
pages. There are specific ways to write code in a css sheet to make it
more organised but I'll have to go through that with you guys on
Thursday.

The HTML page is just the very base, not completed yet, will complete
soon.