A lot of shit just happened
Header and footer have not been completed yet, but a lot of shit is
going on. The images need to be changed to the actual images which are
going to be used.

The background image has been set up in the css however I have not put a
path in yet, so that needs to be written (line 92 of style.css).

Ive scrapped the dedicated css document Mark started off with, it was
full of redundant code which was already in style.css. I recommend the
rest of the html pages go through the same change, especially if the
footer and header are going to be integrated to the other pages...

Will continue later tonight or tomorrow afternoon.